Move out allEventsMetadata() to external file, move it to state...
...update flow
